Import FormatBehaviors from downstream project
Though not currently required by any of JsonRpcMessageSpec, this is the logical
place for the addition of FormatBehaviors. The addition and use of
FormatBehaviors.ordered in FormatBehaviors.decodeError allows the expected
validation errors to be listed in any order. Previously the order of expected
validation errors had to be determined by trial and error, as their order in
JsError objects appears to be non-deterministic.