Change identifier number type from Int to BigDecimal, tweak eitherFormats
Changing to BigDecimal means we support fractional identifiers, which though
discouraged are permitted in JSON-RPC 2.0.

Also in this change (though not used): eitherObjectFormat and eitherValueFormat
now support passing the Formats for their left and right types, enabling
override of any implicits that might be in scope at their callers.

Their read order is swapped too, so first they try to read as the Right type,
and failing that, the Left type. This makes more sense in case of overlapping
JSON representations; we prefer the Right type.