Simplify MessageFormat by removing direct case object support, add objectFormat
As per the updated example, everything that was possible before remains
possible now. The difference is that formats for case objects must be explicitly
created using the new objectFormat method. This actually improves readability on
the user side too, as having to creating a format for objects is consistent with
having to do so for case classes.