Improve spec/test readability
As the name suggests, these are specifications as much as they are tests, and
with that in mind, readability is especially important. So much so that it takes
priority over DRY in this case. Hence the inlining of the helper functions, in
order to remove indirection that made it less than obvious what was happening.