Replace Either[..Notification.., ..Request..] with ..RequestOrNotification..
This eliminates the need to wrap with Right or Left during construction, as the
change in the spec shows, and removes the need to extract during matching.

It also triggers the removal of eitherValueFormat which is now no longer used
(and can be removed even though public because 2.0.0 has not yet been released).
eitherObjectFormat is also removed because it is no longer used as of 4aa65.