Remove "virtual" play-json-rpc module
This was kept after the rename to scala-json-rpc that so users of
play-json-rpc who were not aware of the rename but who were tracking
updates to their play-json-rpc dependency would still get them via the
virtual module (and its dependency on scala-json-rpc). Enough subsequent
releases have now occurred that it will surely have served that purpose.

Also switch to preferred Markdown header syntax in README.